Improves logic of if with SearchStrings
Adding the missing bsc numbers in the else case was not so nice. This
improves the flow a little bit but doesn't change the logic behind it.

SearchStrings might stop with a result < len(array) if it's clear that the
result is not in the array. That's why there also has to be a check if the
result is at the stopped position.